Fix SonarCloud findings (cycle 1)
Split mission_control/consumers.py (status consumers -> status_consumers.py)
and views/_guacamole.py (resolve/generate helpers -> _guacamole_builders.py)
to stay under the 500-line S104 cap; flatten _SecretCache.get to <=3 returns
and add docstrings to the cache-config helpers.
